Initial commit breaking up utils
